+++ packages/varlet-ui/src/avatar/props.ts
@@ -22,6 +22,9 @@ export const props = {
     validator: sizeValidator,
     default: 'normal',
   },
+  alt: {
+    type: String,
+  },
   color: {
     type: String,
   },