+++ packages/varlet-ui/src/fab/Fab.tsx
@@ -1,10 +1,11 @@
 import Button from '../button'
 import Icon from '../icon'
-import { defineComponent, Ref, ref, Teleport, Transition, watch } from 'vue'
+import Drag from '../drag'
+import { defineComponent, Ref, ref, Transition, watch } from 'vue'
 import { useClickOutside } from '@varlet/use'
-import { call, createNamespace, flatFragment, useTeleport, useVModel } from '../utils/components'
+import { isBoolean } from '@varlet/shared'
+import { call, createNamespace, flatFragment, useVModel } from '../utils/components'
 import { toSizeUnit } from '../utils/elements'
-import { toNumber } from '@varlet/shared'
 import { props } from './props'
 
 import '../styles/common.less'
@@ -13,6 +14,7 @@ import '../ripple/ripple.less'
 import '../icon/icon.less'
 import '../loading/loading.less'
 import '../button/button.less'
+import '../drag/drag.less'
 import './fab.less'
 
 const { classes, n } = createNamespace('fab')
@@ -24,7 +26,7 @@ export default defineComponent({
   setup(props, { slots, attrs }) {
     const isActive = useVModel(props, 'active')
     const host: Ref<null | HTMLElement> = ref(null)
-    const { disabled } = useTeleport()
+    const dragRef: Ref<InstanceType<typeof Drag> | null> = ref(null)
 
     const handleClick = (e: Event, value: boolean, childrenLength: number) => {
       e.stopPropagation()
@@ -35,6 +37,11 @@ export default defineComponent({
 
       if (childrenLength === 0) {
         call(props.onClick, isActive.value, e)
+        return
+      }
+
+      // avoid trigger open function after dragging
+      if (dragRef.value?.dragging) {
         return
       }
 
@@ -76,8 +83,8 @@ export default defineComponent({
           type={props.type}
           color={props.color}
           disabled={props.disabled}
-          round
           elevation={props.elevation}
+          round
         >
           <Icon
             var-fab-cover
@@ -88,52 +95,6 @@ export default defineComponent({
             animationClass={n('--trigger-icon-animation')}
           />
         </Button>
-      )
-    }
-
-    const renderFab = () => {
-      const children = flatFragment(call(slots.default) ?? [])
-
-      return (
-        <div
-          class={classes(
-            n(),
-            n(`--position-${props.position}`),
-            n(`--direction-${props.direction}`),
-            [props.fixed, n('--fixed'), n('--absolute')],
-            [props.safeArea, n('--safe-area')]
-          )}
-          style={{
-            zIndex: toNumber(props.zIndex),
-            top: toSizeUnit(props.top),
-            bottom: toSizeUnit(props.bottom),
-            left: toSizeUnit(props.left),
-            right: toSizeUnit(props.right),
-          }}
-          ref={host}
-          onClick={(e) => handleClick(e, !isActive.value, children.length)}
-          onMouseleave={() => handleMouse(false, children.length)}
-          onMouseenter={() => handleMouse(true, children.length)}
-          {...attrs}
-        >
-          <Transition name={n(`--active-transition`)}>{renderTrigger()}</Transition>
-
-          <Transition
-            name={n(`--actions-transition-${props.direction}`)}
-            onAfterEnter={props.onOpened}
-            onAfterLeave={props.onClosed}
-          >
-            <div
-              class={n('actions')}
-              v-show={props.show && isActive.value && children.length}
-              onClick={(e) => e.stopPropagation()}
-            >
-              {children.map((child) => (
-                <div class={n('action')}>{child}</div>
-              ))}
-            </div>
-          </Transition>
-        </div>
       )
     }
 
@@ -151,20 +112,65 @@ export default defineComponent({
       }
     )
 
+    watch(
+      () => [props.position, props.fixed, props.top, props.bottom, props.left, props.right],
+      () => {
+        dragRef.value?.reset()
+      }
+    )
+
     useClickOutside(host, 'click', handleClickOutside)
 
     return () => {
-      const { teleport } = props
+      const children = flatFragment(call(slots.default) ?? [])
 
-      if (teleport) {
-        return (
-          <Teleport to={teleport} disabled={disabled.value}>
-            {renderFab()}
-          </Teleport>
-        )
-      }
+      const dragProps = isBoolean(props.drag) ? {} : props.drag
 
-      return renderFab()
+      return (
+        <Drag
+          ref={dragRef}
+          class={classes(n(`--position-${props.position}`), [!props.fixed, n('--absolute')])}
+          style={{
+            top: toSizeUnit(props.top),
+            bottom: toSizeUnit(props.bottom),
+            left: toSizeUnit(props.left),
+            right: toSizeUnit(props.right),
+          }}
+          zIndex={props.zIndex}
+          teleport={props.teleport}
+          disabled={props.disabled || !props.drag || !props.fixed}
+          direction={dragProps.direction}
+          attraction={dragProps.attraction}
+          boundary={dragProps.boundary}
+          {...attrs}
+        >
+          <div
+            class={classes(n(), n(`--direction-${props.direction}`), [props.safeArea, n('--safe-area')])}
+            ref={host}
+            onClick={(e) => handleClick(e, !isActive.value, children.length)}
+            onMouseleave={() => handleMouse(false, children.length)}
+            onMouseenter={() => handleMouse(true, children.length)}
+          >
+            <Transition name={n(`--active-transition`)}>{renderTrigger()}</Transition>
+
+            <Transition
+              name={n(`--actions-transition-${props.direction}`)}
+              onAfterEnter={props.onOpened}
+              onAfterLeave={props.onClosed}
+            >
+              <div
+                class={n('actions')}
+                v-show={props.show && isActive.value && children.length}
+                onClick={(e) => e.stopPropagation()}
+              >
+                {children.map((child) => (
+                  <div class={n('action')}>{child}</div>
+                ))}
+              </div>
+            </Transition>
+          </div>
+        </Drag>
+      )
     }
   },
 })