+++ packages/varlet-ui/src/fab/example/index.vue
@@ -19,6 +19,7 @@ const direction = ref('top')
 const position = ref('right-bottom')
 const show = ref(true)
 const active = ref(false)
+const drag = ref(false)
 const disabled = ref(false)
 const elevation = ref(2)
 
@@ -69,6 +70,9 @@ watchDarkMode(dark)
     <var-radio checked-value="left"> left </var-radio>
   </var-radio-group>
 
+  <app-type>{{ pack.drag }}</app-type>
+  <var-switch v-model="drag" />
+
   <app-type>{{ pack.disabled }}</app-type>
   <var-switch v-model="disabled" />
 
@@ -92,6 +96,7 @@ watchDarkMode(dark)
     :position="position"
     :direction="direction"
     :trigger="trigger"
+    :drag="drag"
     :disabled="disabled"
     :elevation="elevation"
   >