+++ packages/varlet-ui/src/popup/props.ts
@@ -49,7 +49,7 @@ export const props = {
     default: false,
   },
   teleport: {
-    type: [String, Object] as PropType<TeleportProps['to']>,
+    type: [String, Object, Boolean] as PropType<TeleportProps['to'] | false>,
     default: 'body',
   },
   onOpen: defineListenerProp<() => void>(),